requirements.txt
pyserial
xmodem
